pipeline/requirements.txt
pandas
numpy 
setuptools
mflow
azureml-mlflow
azure-ai-ml
azure-identity
